+++ src/main/java/io/gravitee/policy/trafficshadowing/v3/TrafficShadowingPolicyV3.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright © 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.policy.trafficshadowing.v3;
+
+import io.gravitee.gateway.api.ExecutionContext;
+import io.gravitee.gateway.api.Invoker;
+import io.gravitee.policy.api.PolicyChain;
+import io.gravitee.policy.api.annotations.OnRequest;
+import io.gravitee.policy.trafficshadowing.configuration.TrafficShadowingPolicyConfiguration;
+import io.gravitee.policy.trafficshadowing.v3.invoker.ShadowInvoker;
+
+/**
+ * @author David BRASSELY (david.brassely at graviteesource.com)
+ * @author GraviteeSource Team
+ */
+public class TrafficShadowingPolicyV3 {
+
+    protected final TrafficShadowingPolicyConfiguration configuration;
+
+    public TrafficShadowingPolicyV3(final TrafficShadowingPolicyConfiguration configuration) {
+        this.configuration = configuration;
+    }
+
+    @OnRequest
+    public void onRequest(ExecutionContext context, PolicyChain chain) {
+        // Override the invoker
+        Invoker defaultInvoker = (Invoker) context.getAttribute(ExecutionContext.ATTR_INVOKER);
+        context.setAttribute(ExecutionContext.ATTR_INVOKER, new ShadowInvoker(defaultInvoker, configuration));
+        chain.doNext(context.request(), context.response());
+    }
+}
