+++ src/main/java/io/gravitee/policy/trafficshadowing/v3/invoker/ShadowInvoker.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.policy.trafficshadowing.invoker;
+package io.gravitee.policy.trafficshadowing.v3.invoker;
 
 import static org.springframework.util.StringUtils.hasText;
