+++ src/test/java/io/gravitee/policy/trafficshadowing/v3/invoker/ShadowInvokerTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.policy.trafficshadowing.invoker;
+package io.gravitee.policy.trafficshadowing.v3.invoker;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.AdditionalAnswers.returnsFirstArg;
