+++ src/test/resources/apis/v2/traffic-shadowing-bad-endpoint.json
